Refactor: Update TypeScript Configuration and Introduce Trusted Logos Data
